+++ src/main/java/org/realityforge/revapi/diff/CollectorReporter.java
@@ -1,7 +1,9 @@
 package org.realityforge.revapi.diff;
 
 import java.util.ArrayList;
+import java.util.Comparator;
 import java.util.List;
+import java.util.stream.Collectors;
 import javax.annotation.Nonnull;
 import org.revapi.Report;
 import org.revapi.simple.SimpleReporter;
@@ -11,9 +13,14 @@ public final class CollectorReporter
 {
   private final List<Report> _reports = new ArrayList<>();
 
+  @Nonnull
   List<Report> getReports()
   {
-    return _reports;
+    // Ensure a stable ordering
+    return _reports
+      .stream()
+      .sorted( Comparator.comparing( r -> r.getNewElement() + "-" + r.getOldElement() ) )
+      .collect( Collectors.toList() );
   }
 
   @Override