+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 * Add an option `--expect-no-differences` that will generate a non-zero exit code when differences are
   detected otherwise the tool has been updated to return `0` on success.
+* Ensure that the order of the differences in the report are stable and the order of the keys within a
+  difference are stable. The purpose is to minimize churn when the differences are committed to version
+  control systems.
 
 ### [v0.02](https://github.com/realityforge/revapi-diff/tree/v0.02) (2019-02-09)
 [Full Changelog](https://github.com/realityforge/revapi-diff/compare/v0.01...v0.02)